requirements.txt
crewai
requests
beautifulsoup4
selenium
pillow
groq
langchain
langchain-groq
langchain-litellm
aiofiles
urllib3
lxml
webdriver-manager
python-dotenv
playwright
langchain_community
flask
streamlit
plotly
pandas
nest-asyncio